op-deployer: move InsertPreinstalls from interopgen into op-deployer
